Add cmake-ide to lang/c-c++ layer
cmake-ide nicely orchestrates between cmake, flycheck, company-clang and
RTags by using cmake to generate a compile_commands.json that it uses to
set the appropriate compile flags for flycheck and company-clang. This
nicely centralizes all build related settings where they belong: in the
build system itself.

It comes with a zero need for configuration and usually works just out
of the box. Nevertheless, and recommended as well, build directory and
additional compile flags (...) can also be set through directory local
variables. In the future, we could improve here on the user experience
by providing an automatic and predicable naming scheme for build dirs
that the user can partially influence through a base directory variable.
With that in place, those build directories would be fully automatic and
still reusable by cmake-ide.

Since cmake-ide makes it a breeze to use RTags by automating much of the
necessary setup, I see this as a precursor to proper RTags support which
also requires the rtags lisp package to be installed and required along
with the RTags client/server counterpart (see pull request #2834).